Move path-related classes to the `params` module
